feat(license): ship Nuitka-compiled license_core.so for Linux (LED-1259) — 4.5.11
Closes the docstring-vs-reality gap: license_core.py docstring claimed
"compiled with Nuitka, distributed as a native binary, not readable
Python" but the bundle was shipping plaintext .py source. Customers
could grep their installed file for bypass mechanisms — the surface
that made earlier 4.5.3/4.5.6/4.5.7 bypass leaks discoverable.

What ships in 4.5.11:
- gateway/ai/license_core.cpython-310-x86_64-linux-gnu.so (~370 KB)
  compiled at publish time via Nuitka 2.5.x in the publish workflow
- gateway/ai/license_core.py is now EXCLUDED from the npm bundle
  (.npmignore + package.json files-array)
- license.py shim falls back to Python on ImportError (mac/windows
  customers and dev environments) with a one-line stderr warning

Build pipeline:
- New scripts/build-license-core.sh: nuitka compile + bypass-identifier
  strings-scan + plaintext .py removal from the staging tree
- New scripts/test-license-core-so.sh: isolated tmp-dir compile-and-import
  smoke test with 7 behaviour assertions
- Workflow validate job runs test-license-core-so.sh; publish job runs
  build-license-core.sh before npm pack

Verification (POC):
- Local compile produces 372,456-byte .so
- strings | grep -iE 'JAMSONS|DELIMIT_TEST_MODE|DELIMIT_INTERNAL_LICENSE_KEY'
  → zero matches (vs plaintext .py which contains all three)
- 196 npm tests pass
- delimit-cli-4.5.11.tgz dry-run: 171 files, 823 KB, .so present, .py absent

LIMITATION (filed as follow-up): Linux x86_64 only. mac/windows customers
will hit the ImportError path and fall back to the Python implementation
(degraded but functional). Cross-platform .so/.pyd matrix is a separate
ledger item.
