+++ gateway/ai/license_core.pyi
@@ -1,41 +1,37 @@
 # This file was generated by Nuitka
 
 # Stubs included by default
-from __future__ import annotations
+import time
+import json
+import os
 from pathlib import Path
 import hashlib
-import json
-import time
 
-LICENSE_FILE = Path.home() / '.delimit' / 'license.json'
-USAGE_FILE = Path.home() / '.delimit' / 'usage.json'
-LS_VALIDATE_URL = 'https://api.lemonsqueezy.com/v1/licenses/validate'
-REVALIDATION_INTERVAL = 30 * 86400
-GRACE_PERIOD = 7 * 86400
-HARD_BLOCK = 14 * 86400
-PRO_TOOLS = frozenset({'delimit_gov_health', 'delimit_gov_status', 'delimit_gov_evaluate', 'delimit_gov_policy', 'delimit_gov_run', 'delimit_gov_verify', 'delimit_deploy_plan', 'delimit_deploy_build', 'delimit_deploy_publish', 'delimit_deploy_verify', 'delimit_deploy_rollback', 'delimit_deploy_site', 'delimit_deploy_npm', 'delimit_memory_store', 'delimit_memory_search', 'delimit_memory_recent', 'delimit_vault_search', 'delimit_vault_snapshot', 'delimit_vault_health', 'delimit_evidence_collect', 'delimit_evidence_verify', 'delimit_deliberate', 'delimit_models', 'delimit_obs_metrics', 'delimit_obs_logs', 'delimit_obs_status', 'delimit_release_plan', 'delimit_release_status', 'delimit_release_sync', 'delimit_cost_analyze', 'delimit_cost_optimize', 'delimit_cost_alert'})
-FREE_TRIAL_LIMITS = {'delimit_deliberate': 3}
+PRO_TOOLS = frozenset({'delimit_gov_evaluate', 'delimit_gov_policy', 'delimit_gov_run', 'delimit_gov_verify', 'delimit_os_plan', 'delimit_os_status', 'delimit_os_gates', 'delimit_deploy_plan', 'delimit_deploy_build', 'delimit_deploy_publish', 'delimit_deploy_verify', 'delimit_deploy_rollback', 'delimit_deploy_status', 'delimit_deploy_site', 'delimit_deploy_npm', 'delimit_memory_search', 'delimit_vault_search', 'delimit_vault_snapshot', 'delimit_vault_health', 'delimit_evidence_collect', 'delimit_evidence_verify', 'delimit_deliberate', 'delimit_models', 'delimit_obs_metrics', 'delimit_obs_logs', 'delimit_obs_status', 'delimit_release_plan', 'delimit_release_status', 'delimit_release_sync', 'delimit_cost_analyze', 'delimit_cost_optimize', 'delimit_cost_alert', 'delimit_social_post', 'delimit_social_generate', 'delimit_social_history', 'delimit_screen_record', 'delimit_screenshot', 'delimit_notify', 'delimit_agent_dispatch', 'delimit_agent_status', 'delimit_agent_complete', 'delimit_agent_handoff', 'delimit_executor'})
+def needs_revalidation(data: dict) -> bool:
+    ...
+def revalidate_license(data: dict) -> dict:
+    ...
+def is_license_valid(data: dict) -> bool:
+    ...
+def _write_license(data: dict) -> None:
+    ...
+def _call_lemon_squeezy(data: dict) -> bool | None:
+    ...
 def load_license() -> dict:
     ...
-
 def check_premium() -> bool:
     ...
-
 def gate_tool(tool_name: str) -> dict | None:
     ...
-
 def activate(key: str) -> dict:
     ...
-
 def _revalidate(data: dict) -> dict:
     ...
-
 def _get_monthly_usage(tool_name: str) -> int:
     ...
-
 def _increment_usage(tool_name: str) -> int:
     ...
-
 
 __name__ = ...
 
@@ -44,7 +40,9 @@ __name__ = ...
 # Modules used internally, to allow implicit dependencies to be seen:
 import hashlib
 import json
+import os
 import time
 import pathlib
 import urllib
 import urllib.request
+import re
